Move test to correct location
